Vote-Button Ajax-fähig gemacht

+++ delete.php
@@ -1,11 +1,14 @@
 <?php
+include_once ('session_check.php');
 
 $contentID = (int)$_GET["contentID"];
+$userID = $_SESSION['userid'];
+
 
 try {
     include_once("userdata.php");
     $db = new PDO($dsn, $dbuser, $dbpass);
-    $sql = "DELETE FROM content_txt WHERE contentID=$contentID";
+    $sql = "DELETE FROM content_txt WHERE contentID=$contentID AND userid=$userID";
     $db->prepare($sql)->execute();
     $db = null;
 } catch (PDOException $e) {
